refactor(server): extract season and battle-snapshot types into persistence/types.ts (part of #1559) by claude
Moves SeasonSnapshot, SeasonListOptions, SeasonCloseSummary, LeaderboardSeasonArchiveEntry,
PlayerReferralClaimResult, BattleSnapshot* family, SnapshotRetentionPolicy, and
PlayerNameHistoryRetentionPolicy into apps/server/src/persistence/types.ts.

The types are re-exported from persistence.ts to preserve the public import surface.
No runtime behavior changes. Refs #1559.
